+++ integrated_channels/moodle/client.py
@@ -74,14 +74,6 @@ def moodle_request_wrapper(method):
             # but we have nothing else to go on
             if body == 0:
                 if method.__name__ == "_wrapped_create_course_completion" and response.status_code == 200:
-                    LOGGER.info(
-                        'Integer Response for Moodle Course Completion'
-                        f'with data kwargs={kwargs} and args={args} '
-                        f' response: {response} '
-                        f'Status Code: {response.status_code}, '
-                        f'Text: {response.text}, '
-                        f'Headers: {response.headers}, '
-                    )
                     return MoodleResponse(status_code=200, text='')
                 return 200, ''
             raise ClientError('Moodle API Grade Update failed with int code: {code}'.format(code=body), 500)